requirements.txt
Django==4.2.7
django-filter==24.1
djangorestframework==3.14.0
geopy==2.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0